requirements.txt
cmake==3.18.4
wheel
numpy>1.2
hydra-core==1.1.1
hydra-colorlog
matplotlib~=3.6.1
opencv-python~=4.6.0.66
omegaconf~=2.1.2
plotly~=5.10.0
ftfy~=6.1.1
pyhash~=0.9.3
pytorch-lightning==1.5.9
torch==1.10.1
torchvision~=0.11.2
MulticoreTSNE~=0.1
gitpython
scipy~=1.7.2
sentence-transformers
pytorch3d
gym~=0.26.2
moviepy
tqdm~=4.64.1
termcolor~=2.0.1
wandb~=0.13.9

regex~=2022.9.13
Pillow~=9.2.0
pybullet~=3.2.5
trimesh~=3.15.5
urdfpy~=0.0.22
pyrender~=0.1.45
setuptools